Tags that do not fit any of the skills in the list now get a disabled checkbox
Added the function disableUnavailableCheckBoxes that checks for every tag
if this tag still appears in one of the skills from the fitting skills. If
not, disables the checkbox. If the tag appears in one or more skills but
the button is disabled the funtion enables it again.
Also added the getPrime(String criterium) function in the SkillFilter in
case you only need the prime of one tag. Before you had to first construct
an ArrayList to obtain the prime of one tag.