Added a table with all relevant information about each skill that is displayed in the result list. The Panes in this list are now saved in the skill class and are not created each time a checkbox changes its state any more.
The Information for the tables for each skill is stored in the ini file
and is now also read at the beginning by the InitReader. While reading the
complete Pane that is displayed in the result list is now created and
stored in the skill class. This should reduce the loading time but will
increase RAM usage.