Added a Class for writing Error messages to a file. Also fixed the size of the result list.
Added the ErrorFileWriter to be able to write errors to a file instead of
to the commandline. I could have used something like log4j but I decided
to keep the programm as small as possible and didnt want to include the
whole log4j library.
Gave the ScrollPane a prefered size, it now is big enough to display all
the skills (curses, cough) in their entirety without having to scroll.
Also gave all important numeric values their own private final int
variable on top of the class.